Add files via upload
This is a Node.JS wrapper that imports patients utilizing a CSV file into the sandbox tenant. This wrapper runs the information provided through a series of API calls within Health Gorilla P360. Information in the CSV file need to be updated as well bearer token for the tenant needs to be placed. All dependencies must also be installed for this node wrapper to run properly. 

This Wrapper is to be used as an example and starting point for developers to automate the API calls for patient 360.